requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
torch==2.2.0
torchvision==0.17.0
ultralytics==8.2.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.26.3
Pillow==10.2.0
python-multipart==0.0.9
cloudinary==1.37.0
python-jose==3.3.0